Move `get_service_endorsements` to public header
