+++ cmake/common.cmake
@@ -184,6 +184,7 @@ set(CCF_ENDPOINTS_SOURCES
     ${CCF_DIR}/src/indexing/strategies/seqnos_by_key_in_memory.cpp
     ${CCF_DIR}/src/indexing/strategies/visit_each_entry_in_map.cpp
     ${CCF_DIR}/src/node/historical_queries_adapter.cpp
+    ${CCF_DIR}/src/node/historical_queries_utils.cpp
     ${CCF_DIR}/src/node/receipt.cpp
 )
