+++ include/ccf/network_identity_interface.h
