+++ src/node/historical_queries_adapter.cpp
@@ -3,10 +3,10 @@
 
 #include "ccf/historical_queries_adapter.h"
 
+#include "ccf/historical_queries_utils.h"
 #include "ccf/rpc_context.h"
 #include "ccf/service/tables/service.h"
 #include "kv/kv_types.h"
-#include "node/historical_queries_utils.h"
 #include "node/rpc/network_identity_subsystem.h"
 #include "node/tx_receipt_impl.h"
 
@@ -319,7 +319,7 @@ namespace ccf::historical
       if (
         historical_state == nullptr ||
         (!get_service_endorsements(
-          args, historical_state, state_cache, network_identity_subsystem)))
+          args.tx, historical_state, state_cache, network_identity_subsystem)))
       {
         args.rpc_ctx->set_response_status(HTTP_STATUS_ACCEPTED);
         constexpr size_t retry_after_seconds = 3;
