+++ src/node/rpc/network_identity_subsystem.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/network_identity_interface.h"
 #include "node/identity.h"
-#include "node/rpc/network_identity_interface.h"
 #include "node/rpc/node_interface.h"
 
 namespace ccf
